+++ test_unstructured/partition/test_email.py
@@ -26,7 +26,8 @@ from unstructured.partition.email import (
     partition_email_header,
 )
 
-DIRECTORY = pathlib.Path(__file__).parent.resolve()
+FILE_DIRECTORY = pathlib.Path(__file__).parent.resolve()
+EXAMPLE_DOCS_DIRECTORY = os.path.join(FILE_DIRECTORY, "..", "..", "example-docs", "eml")
 
 
 EXPECTED_OUTPUT = [
@@ -100,25 +101,37 @@ ATTACH_EXPECTED_OUTPUT = [
 
 
 def test_partition_email_from_filename():
-    filename = os.path.join(DIRECTORY, "..", "..", "example-docs", "fake-email.eml")
+    filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "fake-email.eml")
     elements = partition_email(filename=filename)
     assert len(elements) > 0
     assert elements == EXPECTED_OUTPUT
 
 
 @pytest.mark.parametrize(
-    "filename",
-    ["fake-email-utf-16.eml", "fake-email-utf-16-be.eml", "fake-email-utf-16-le.eml"],
+    ("filename", "expected_output"),
+    [
+        ("fake-email-utf-16.eml", EXPECTED_OUTPUT),
+        ("fake-email-utf-16-be.eml", EXPECTED_OUTPUT),
+        ("fake-email-utf-16-le.eml", EXPECTED_OUTPUT),
+        ("email-no-utf8-2008-07-16.062410.eml", None),
+        ("email-no-utf8-2014-03-17.111517.eml", None),
+        ("email-replace-mime-encodings-error-1.eml", None),
+        ("email-replace-mime-encodings-error-2.eml", None),
+        ("email-replace-mime-encodings-error-3.eml", None),
+        ("email-replace-mime-encodings-error-4.eml", None),
+        ("email-replace-mime-encodings-error-5.eml", None),
+    ],
 )
-def test_partition_email_from_filename_default_encoding(filename):
-    filename = os.path.join(DIRECTORY, "..", "..", "example-docs", filename)
+def test_partition_email_from_filename_default_encoding(filename, expected_output):
+    filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, filename)
     elements = partition_email(filename=filename)
     assert len(elements) > 0
-    assert elements == EXPECTED_OUTPUT
+    if expected_output:
+        assert elements == expected_output
 
 
 def test_partition_email_from_file():
-    filename = os.path.join(DIRECTORY, "..", "..", "example-docs", "fake-email.eml")
+    filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "fake-email.eml")
     with open(filename) as f:
         elements = partition_email(file=f)
     assert len(elements) > 0
@@ -126,19 +139,31 @@ def test_partition_email_from_file():
 
 
 @pytest.mark.parametrize(
-    "filename",
-    ["fake-email-utf-16.eml", "fake-email-utf-16-be.eml", "fake-email-utf-16-le.eml"],
+    ("filename", "expected_output"),
+    [
+        ("fake-email-utf-16.eml", EXPECTED_OUTPUT),
+        ("fake-email-utf-16-be.eml", EXPECTED_OUTPUT),
+        ("fake-email-utf-16-le.eml", EXPECTED_OUTPUT),
+        ("email-no-utf8-2008-07-16.062410.eml", None),
+        ("email-no-utf8-2014-03-17.111517.eml", None),
+        ("email-replace-mime-encodings-error-1.eml", None),
+        ("email-replace-mime-encodings-error-2.eml", None),
+        ("email-replace-mime-encodings-error-3.eml", None),
+        ("email-replace-mime-encodings-error-4.eml", None),
+        ("email-replace-mime-encodings-error-5.eml", None),
+    ],
 )
-def test_partition_email_from_file_default_encoding(filename):
-    filename = os.path.join(DIRECTORY, "..", "..", "example-docs", filename)
+def test_partition_email_from_file_default_encoding(filename, expected_output):
+    filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, filename)
     with open(filename) as f:
         elements = partition_email(file=f)
     assert len(elements) > 0
-    assert elements == EXPECTED_OUTPUT
+    if expected_output:
+        assert elements == expected_output
 
 
 def test_partition_email_from_file_rb():
-    filename = os.path.join(DIRECTORY, "..", "..", "example-docs", "fake-email.eml")
+    filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "fake-email.eml")
     with open(filename, "rb") as f:
         elements = partition_email(file=f)
     assert len(elements) > 0
@@ -146,19 +171,31 @@ def test_partition_email_from_file_rb():
 
 
 @pytest.mark.parametrize(
-    "filename",
-    ["fake-email-utf-16.eml", "fake-email-utf-16-be.eml", "fake-email-utf-16-le.eml"],
+    ("filename", "expected_output"),
+    [
+        ("fake-email-utf-16.eml", EXPECTED_OUTPUT),
+        ("fake-email-utf-16-be.eml", EXPECTED_OUTPUT),
+        ("fake-email-utf-16-le.eml", EXPECTED_OUTPUT),
+        ("email-no-utf8-2008-07-16.062410.eml", None),
+        ("email-no-utf8-2014-03-17.111517.eml", None),
+        ("email-replace-mime-encodings-error-1.eml", None),
+        ("email-replace-mime-encodings-error-2.eml", None),
+        ("email-replace-mime-encodings-error-3.eml", None),
+        ("email-replace-mime-encodings-error-4.eml", None),
+        ("email-replace-mime-encodings-error-5.eml", None),
+    ],
 )
-def test_partition_email_from_file_rb_default_encoding(filename):
-    filename = os.path.join(DIRECTORY, "..", "..", "example-docs", filename)
+def test_partition_email_from_file_rb_default_encoding(filename, expected_output):
+    filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, filename)
     with open(filename, "rb") as f:
         elements = partition_email(file=f)
     assert len(elements) > 0
-    assert elements == EXPECTED_OUTPUT
+    if expected_output:
+        assert elements == expected_output
 
 
 def test_partition_email_from_text_file():
-    filename = os.path.join(DIRECTORY, "..", "..", "example-docs", "fake-email.txt")
+    filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "fake-email.txt")
     with open(filename) as f:
         elements = partition_email(file=f, content_source="text/plain")
     assert len(elements) > 0
@@ -166,7 +203,7 @@ def test_partition_email_from_text_file():
 
 
 def test_partition_email_from_text_file_with_headers():
-    filename = os.path.join(DIRECTORY, "..", "..", "example-docs", "fake-email.txt")
+    filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "fake-email.txt")
     with open(filename) as f:
         elements = partition_email(file=f, content_source="text/plain", include_headers=True)
     assert len(elements) > 0
@@ -174,7 +211,7 @@ def test_partition_email_from_text_file_with_headers():
 
 
 def test_partition_email_from_text():
-    filename = os.path.join(DIRECTORY, "..", "..", "example-docs", "fake-email.eml")
+    filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "fake-email.eml")
     with open(filename) as f:
         text = f.read()
     elements = partition_email(text=text)
@@ -187,14 +224,14 @@ def test_partition_email_from_text_work_with_empty_string():
 
 
 def test_partition_email_from_filename_with_embedded_image():
-    filename = os.path.join(DIRECTORY, "..", "..", "example-docs", "fake-email-image-embedded.eml")
+    filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "fake-email-image-embedded.eml")
     elements = partition_email(filename=filename, content_source="text/plain")
     assert len(elements) > 0
     assert elements == IMAGE_EXPECTED_OUTPUT
 
 
 def test_partition_email_header():
-    filename = os.path.join(DIRECTORY, "..", "..", "example-docs", "fake-email-header.eml")
+    filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "fake-email-header.eml")
     with open(filename) as f:
         msg = email.message_from_file(f)
     elements = partition_email_header(msg)
@@ -203,7 +240,7 @@ def test_partition_email_header():
 
 
 def test_partition_email_has_metadata():
-    filename = os.path.join(DIRECTORY, "..", "..", "example-docs", "fake-email-header.eml")
+    filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "fake-email-header.eml")
     elements = partition_email(filename=filename)
     assert len(elements) > 0
     assert (
@@ -225,7 +262,7 @@ def test_partition_email_has_metadata():
 
 
 def test_extract_email_text_matches_html():
-    filename = os.path.join(DIRECTORY, "..", "..", "example-docs", "fake-email-attachment.eml")
+    filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "fake-email-attachment.eml")
     elements_from_text = partition_email(filename=filename, content_source="text/plain")
     elements_from_html = partition_email(filename=filename, content_source="text/html")
 
@@ -237,7 +274,7 @@ def test_extract_email_text_matches_html():
 
 
 def test_extract_attachment_info():
-    filename = os.path.join(DIRECTORY, "..", "..", "example-docs", "fake-email-attachment.eml")
+    filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "fake-email-attachment.eml")
     with open(filename) as f:
         msg = email.message_from_file(f)
     attachment_info = extract_attachment_info(msg)
@@ -251,7 +288,7 @@ def test_partition_email_raises_with_none_specified():
 
 
 def test_partition_email_raises_with_too_many_specified():
-    filename = os.path.join(DIRECTORY, "..", "..", "example-docs", "fake-email.eml")
+    filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "fake-email.eml")
     with open(filename) as f:
         text = f.read()
 
@@ -260,13 +297,13 @@ def test_partition_email_raises_with_too_many_specified():
 
 
 def test_partition_email_raises_with_invalid_content_type():
-    filename = os.path.join(DIRECTORY, "..", "..", "example-docs", "fake-email.eml")
+    filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "fake-email.eml")
     with pytest.raises(ValueError):
         partition_email(filename=filename, content_source="application/json")
 
 
 def test_partition_email_processes_fake_email_with_header():
-    filename = os.path.join(DIRECTORY, "..", "..", "example-docs", "fake-email-header.eml")
+    filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "fake-email-header.eml")
     elements = partition_email(filename=filename)
     assert len(elements) > 0
 
@@ -283,3 +320,9 @@ def test_partition_email_processes_fake_email_with_header():
 def test_convert_to_iso_8601(time, expected):
     iso_time = convert_to_iso_8601(time)
     assert iso_time == expected
+
+
+def test_partition_email_raises_with_no_html_content():
+    filename = os.path.join(EXAMPLE_DOCS_DIRECTORY, "email-no-html-content-1.eml")
+    with pytest.raises(ValueError):
+        partition_email(filename=filename)